requirement.txt
torch
streamlit
torchvision
matplotlib
Pillow
